Compatibility with Numpy v1.24 (#208)
* compatibility with numpy 1.24

* update CI to newest action, generic gfortran, and python 3.9

* install gfortran on macs

* change slashes on windows?

* revert slash and shell changes

* more path fix

* trying quotes

* try bash again

* specify source dir

* trial and error. . .

* trial and error. . .

* try a different shell?

* trial and error

* trial and error

* trial and error

+++ ROSCO_toolbox/ofTools/fast_io/read_fast_input.py
@@ -802,9 +802,9 @@ class FASTInputFile(File):
                             self.addKeyVal('nDOF',int(l.split(':')[1]))
                             nDOFCommon=self['nDOF']
                         elif l.find('!time increment')==0:
-                            self.addKeyVal('dt',np.float(l.split(':')[1]))
+                            self.addKeyVal('dt',np.float_(l.split(':')[1]))
                         elif l.find('!total simulation time')==0:
-                            self.addKeyVal('T',np.float(l.split(':')[1]))
+                            self.addKeyVal('T',np.float_(l.split(':')[1]))
                     else:
                         raise BrokenFormatError('Unexcepted content found on line {}'.format(i))
                 i+=1
@@ -855,9 +855,9 @@ class FASTInputFile(File):
                 nTabLines=0
                 while 14+nTabLines<len(lines) and  len(lines[14+nTabLines].strip())>0 :
                     nTabLines +=1
-                #data = np.array([lines[i].strip().split() for i in range(14,len(lines)) if len(lines[i])>0]).astype(np.float)
-                #data = np.array([lines[i].strip().split() for i in takewhile(lambda x: len(lines[i].strip())>0, range(14,len(lines)-1))]).astype(np.float)
-                data = np.array([lines[i].strip().split() for i in range(14,nTabLines+14)]).astype(np.float)
+                #data = np.array([lines[i].strip().split() for i in range(14,len(lines)) if len(lines[i])>0]).astype(np.float_)
+                #data = np.array([lines[i].strip().split() for i in takewhile(lambda x: len(lines[i].strip())>0, range(14,len(lines)-1))]).astype(np.float_)
+                data = np.array([lines[i].strip().split() for i in range(14,nTabLines+14)]).astype(np.float_)
                 #print(data)
                 d = getDict()
                 d['label']     = 'Polar'
@@ -883,9 +883,9 @@ class FASTInputFile(File):
             for j in range(nStations):
                 M[j,0]=float(lines[i]); i+=1;
                 LL = lines[i:i+6]
-                M[j,1:37]=np.array((' '.join(lines[i:i+6])).split()).astype(np.float)
+                M[j,1:37]=np.array((' '.join(lines[i:i+6])).split()).astype(np.float_)
                 i+=7
-                M[j,37:]=np.array((' '.join(lines[i:i+6])).split()).astype(np.float)
+                M[j,37:]=np.array((' '.join(lines[i:i+6])).split()).astype(np.float_)
                 i+=7
         except: 
             raise WrongFormatError('An error occured while reading section {}/{}'.format(j+1,nStations))
